requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
qdrant-client>=1.7.0
httpx>=0.26.0
python-multipart>=0.0.6
pyyaml>=6.0.0
sse-starlette>=1.6.0
